2019: Day 14: Added part two

+++ 2019/src/main/java/com/augustoccesar/adventofcode/day14/Order.java
@@ -1,5 +1,5 @@
 package com.augustoccesar.adventofcode.day14;
 
-public record Order(int amount, String chemical) {
+public record Order(long amount, String chemical) {
 
 }
