+++ 2019/src/main/java/com/augustoccesar/adventofcode/day14/Reaction.java
@@ -14,19 +14,19 @@ class Reaction {
   private static final Pattern REACTION_REGEX = Pattern.compile("(\\d+)\\s(\\w+)");
 
   @Getter private final List<String> inputChemicals;
-  @Getter private final List<Integer> inputAmounts;
+  @Getter private final List<Long> inputAmounts;
 
   @Getter private final String outputChemical;
-  @Getter private final int outputAmount;
+  @Getter private final long outputAmount;
 
   public static Reaction from(final String reactionInstruction) {
     final Matcher matcher = REACTION_REGEX.matcher(reactionInstruction);
 
     final List<String> chemicals = new ArrayList<>();
-    final List<Integer> amounts = new ArrayList<>();
+    final List<Long> amounts = new ArrayList<>();
     while (matcher.find()) {
       chemicals.add(matcher.group(2));
-      amounts.add(Integer.parseInt(matcher.group(1)));
+      amounts.add(Long.parseLong(matcher.group(1)));
     }
 
     final int lastIdx = chemicals.size() - 1;
